requirements.txt
Telethon == 1.24.0
python_decouple == 3.5
